% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{dist_spec_to_julia}
\alias{dist_spec_to_julia}
\title{Conversion layer between R and Julia}
\usage{
dist_spec_to_julia(dist)
}
\arguments{
\item{dist}{A \code{dist_spec} object}
}
\value{
A JuliaCall proxy object representing the Julia distribution
}
\description{
Internal functions for converting R objects (dist_spec,
*_opts) to their Julia equivalents and converting Julia results back to
R data.tables.
}
\keyword{Convert}
\keyword{Julia}
\keyword{R}
\keyword{a}
\keyword{an}
\keyword{dist_spec}
\keyword{distribution}
\keyword{internal}
\keyword{to}
