% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{dist_spec_uncertain_to_julia}
\alias{dist_spec_uncertain_to_julia}
\title{Convert an uncertain dist_spec to a Julia UncertainDistribution}
\usage{
dist_spec_uncertain_to_julia(dist, family, params, uncertain)
}
\arguments{
\item{dist}{The dist_spec object}

\item{family}{Character, the distribution family}

\item{params}{Named list of parameters}

\item{uncertain}{Logical vector indicating which params are uncertain}
}
\value{
A JuliaCall proxy object
}
\description{
Convert an uncertain dist_spec to a Julia UncertainDistribution
}
\keyword{internal}
