% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\docType{data}
\name{.julia_env}
\alias{.julia_env}
\title{Julia backend initialisation}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.julia_env
}
\description{
Internal functions for managing the Julia backend used by
EpiNow2 for Bayesian inference via Turing.jl.
}
\keyword{internal}
