% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_delays}
\alias{extract_delays}
\title{Extract samples from all delay parameters}
\usage{
extract_delays(samples, args)
}
\arguments{
\item{samples}{Extracted model samples}

\item{args}{Model arguments list}
}
\value{
A \verb{<data.table>} with variable, sample, value columns
}
\description{
Extract samples from all delay parameters
}
\keyword{internal}
