% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_inits}
\alias{extract_inits}
\title{Generate initial conditions from a fit}
\usage{
extract_inits(fit, current_inits, exclude_list = NULL, samples = 50)
}
\arguments{
\item{fit}{A Julia result object.}

\item{current_inits}{A function returning initial conditions.}

\item{exclude_list}{Parameters to exclude.}

\item{samples}{Number of posterior samples.}
}
\value{
A function that returns initial conditions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Extracts posterior samples to use as initial conditions for a new fit.
}
