% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_latent_state}
\alias{extract_latent_state}
\title{Extract samples for a latent state from a model}
\usage{
extract_latent_state(param, samples, dates)
}
\arguments{
\item{param}{Character string indicating the latent state to extract}

\item{samples}{Extracted model samples (list of arrays)}

\item{dates}{A vector of dates for the time dimension}
}
\value{
A \verb{<data.frame>} with time, date, sample, and value columns
}
\description{
Extract samples for a latent state from a model
}
\keyword{internal}
