% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract samples from all parameters}
\usage{
extract_parameters(samples, args)
}
\arguments{
\item{samples}{Extracted model samples}

\item{args}{Model arguments list}
}
\value{
A \verb{<data.table>} with variable, sample, value columns
}
\description{
Extract samples from all parameters
}
\keyword{internal}
