% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_stan_param}
\alias{extract_stan_param}
\title{Extract a parameter summary from a fitted model}
\usage{
extract_stan_param(
  fit,
  params = NULL,
  CrIs = c(0.2, 0.5, 0.9),
  var_names = FALSE
)
}
\arguments{
\item{fit}{A Julia result object.}

\item{params}{A character vector of parameters to extract.}

\item{CrIs}{Numeric vector of credible intervals to calculate.}

\item{var_names}{Logical defaults to \code{FALSE}. Should variables be named.}
}
\value{
A \verb{<data.table>} summarising parameter posteriors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Extracts summarised parameter posteriors from a fitted model.
}
