% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{julia_dist_to_r}
\alias{julia_dist_to_r}
\title{Convert a Julia distribution back to an R dist_spec}
\usage{
julia_dist_to_r(jdist)
}
\arguments{
\item{jdist}{A JuliaCall proxy for a Julia distribution}
}
\value{
A \code{dist_spec} object
}
\description{
Convert a Julia distribution back to an R dist_spec
}
\keyword{internal}
