% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{julia_parameters_to_r}
\alias{julia_parameters_to_r}
\title{Convert Julia parameters to R dist_spec list}
\usage{
julia_parameters_to_r(julia_result)
}
\arguments{
\item{julia_result}{JuliaCall proxy for the Julia result}
}
\value{
A named list of \code{dist_spec} objects
}
\description{
Convert Julia parameters to R dist_spec list
}
\keyword{internal}
