% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{julia_samples_to_r}
\alias{julia_samples_to_r}
\title{Convert Julia EstimateInfectionsResult samples to R data.table}
\usage{
julia_samples_to_r(julia_result, observations, horizon = 0, seeding_time = 0)
}
\arguments{
\item{julia_result}{JuliaCall proxy for the Julia result}

\item{observations}{The original R observations data.frame}

\item{horizon}{Integer, the forecast horizon}

\item{seeding_time}{Integer, the seeding time}
}
\value{
A \code{data.table} with columns: variable, time, date, sample, value,
strat, type
}
\description{
Convert Julia EstimateInfectionsResult samples to R data.table
}
\keyword{internal}
