% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{julia_summary_to_r}
\alias{julia_summary_to_r}
\title{Convert Julia summary DataFrame to R data.table}
\usage{
julia_summary_to_r(julia_df)
}
\arguments{
\item{julia_df}{JuliaCall proxy for a Julia DataFrame with summary stats}
}
\value{
A \code{data.table} with columns: date, mean, median, sd, and CrI columns
}
\description{
Convert Julia summary DataFrame to R data.table
}
\keyword{internal}
