% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_backcalc_opts_to_julia}
\alias{r_backcalc_opts_to_julia}
\title{Convert R backcalc_opts to Julia BackcalcOpts}
\usage{
r_backcalc_opts_to_julia(opts)
}
\arguments{
\item{opts}{A \code{backcalc_opts} object}
}
\value{
JuliaCall proxy for Julia BackcalcOpts
}
\description{
Convert R backcalc_opts to Julia BackcalcOpts
}
\keyword{internal}
