% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_data_to_julia}
\alias{r_data_to_julia}
\title{Convert R data.frame to Julia DataFrame}
\usage{
r_data_to_julia(data)
}
\arguments{
\item{data}{A data.frame with columns \code{date} and \code{confirm}}
}
\value{
JuliaCall proxy for a Julia DataFrame
}
\description{
Convert R data.frame to Julia DataFrame
}
\keyword{internal}
