% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_delay_opts_to_julia}
\alias{r_delay_opts_to_julia}
\title{Convert R delay_opts to Julia DelayOpts}
\usage{
r_delay_opts_to_julia(opts)
}
\arguments{
\item{opts}{A \code{delay_opts} object}
}
\value{
JuliaCall proxy for Julia DelayOpts
}
\description{
Convert R delay_opts to Julia DelayOpts
}
\keyword{internal}
