% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_forecast_opts_to_julia}
\alias{r_forecast_opts_to_julia}
\title{Convert R forecast_opts to Julia ForecastOpts}
\usage{
r_forecast_opts_to_julia(opts)
}
\arguments{
\item{opts}{A \code{forecast_opts} object}
}
\value{
JuliaCall proxy for Julia ForecastOpts
}
\description{
Convert R forecast_opts to Julia ForecastOpts
}
\keyword{internal}
