% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_gp_opts_to_julia}
\alias{r_gp_opts_to_julia}
\title{Convert R gp_opts to Julia GPOpts}
\usage{
r_gp_opts_to_julia(opts)
}
\arguments{
\item{opts}{A \code{gp_opts} object or NULL}
}
\value{
JuliaCall proxy for Julia GPOpts, or NULL
}
\description{
Convert R gp_opts to Julia GPOpts
}
\keyword{internal}
