% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_gt_opts_to_julia}
\alias{r_gt_opts_to_julia}
\title{Convert R gt_opts to Julia GTOpts}
\usage{
r_gt_opts_to_julia(opts)
}
\arguments{
\item{opts}{A \code{generation_time_opts} object}
}
\value{
JuliaCall proxy for Julia GTOpts
}
\description{
Convert R gt_opts to Julia GTOpts
}
\keyword{internal}
