% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_obs_opts_to_julia}
\alias{r_obs_opts_to_julia}
\title{Convert R obs_opts to Julia ObsOpts}
\usage{
r_obs_opts_to_julia(opts)
}
\arguments{
\item{opts}{An \code{obs_opts} object}
}
\value{
JuliaCall proxy for Julia ObsOpts
}
\description{
Convert R obs_opts to Julia ObsOpts
}
\keyword{internal}
