% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_rt_opts_to_julia}
\alias{r_rt_opts_to_julia}
\title{Convert R rt_opts to Julia RtOpts}
\usage{
r_rt_opts_to_julia(opts)
}
\arguments{
\item{opts}{An \code{rt_opts} object or NULL}
}
\value{
JuliaCall proxy for Julia RtOpts, or NULL
}
\description{
Convert R rt_opts to Julia RtOpts
}
\keyword{internal}
