% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_secondary_opts_to_julia}
\alias{r_secondary_opts_to_julia}
\title{Convert R secondary_opts to Julia SecondaryOpts}
\usage{
r_secondary_opts_to_julia(opts)
}
\arguments{
\item{opts}{A \code{secondary_opts} object}
}
\value{
JuliaCall proxy for Julia SecondaryOpts
}
\description{
Convert R secondary_opts to Julia SecondaryOpts
}
\keyword{internal}
