% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{r_trunc_opts_to_julia}
\alias{r_trunc_opts_to_julia}
\title{Convert R trunc_opts to Julia TruncOpts}
\usage{
r_trunc_opts_to_julia(opts)
}
\arguments{
\item{opts}{A \code{trunc_opts} object}
}
\value{
JuliaCall proxy for Julia TruncOpts
}
\description{
Convert R trunc_opts to Julia TruncOpts
}
\keyword{internal}
